# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gerenciamento de Consultas",
        "description": "API para gerenciamento de consultas médicas",
        "contact": {
            "email": "seu-email@exemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/consultas": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Lista todas as consultas",
                "description": "Retorna a lista de todas as consultas agendadas",
                "operationId": "aa009f5c5eed4bebbe800f2f9354abc0",
                "responses": {
                    "200": {
                        "description": "Lista de consultas recuperada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de consultas recuperada com sucesso"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "medico_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "paciente_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "data_hora": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-02-01 14:30:00"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "agendada"
                                                    },
                                                    "medico": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nome": {
                                                                "type": "string",
                                                                "example": "Dr. João Silva"
                                                            },
                                                            "especialidade": {
                                                                "type": "string",
                                                                "example": "Cardiologia"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "paciente": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nome": {
                                                                "type": "string",
                                                                "example": "Maria Santos"
                                                            },
                                                            "telefone": {
                                                                "type": "string",
                                                                "example": "(11) 98765-4321"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/medicos/consulta": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Agenda uma nova consulta",
                "description": "Cria um novo agendamento de consulta",
                "operationId": "240473d85b77f35cd648ddb681244e18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "medico_id",
                                    "paciente_id",
                                    "data_hora"
                                ],
                                "properties": {
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-02-01 14:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta agendada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Consulta agendada com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "medico_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "paciente_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data_hora": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-02-01 14:30:00"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "agendada"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/consultas/{id}": {
            "put": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Atualiza uma consulta existente",
                "description": "Atualiza os dados de uma consulta específica",
                "operationId": "17e66b87ea54847225295804b69d6c12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data_hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-02-01 14:30:00"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "realizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consulta atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Consulta atualizada com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "medico_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "paciente_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data_hora": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-02-01 14:30:00"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "realizada"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}